requirements.txt
Django==5.2.1
psycopg2-binary>=2.9
python-telegram-bot==20.7
python-dotenv>=1.0
Pillow>=10.0